feat: update import of real clinical data
